--- a/src/Collection.php
+++ b/src/Collection.php
@@ -3,7 +3,6 @@
 namespace DusanKasan\Knapsack;
 
 use DusanKasan\Knapsack\Exceptions\InvalidArgument;
-use Iterator;
 use IteratorAggregate;
 use RecursiveArrayIterator;
 use Traversable;