@DusanKasan 2017-04-13
Hi, I'm glad you like Knapsack. First of all thank you for putting the time and effort into this PR. It seems like a great idea and the only BC i see is indeed dropping the Iterator implementation.

The code you submitted looks good except that [the callable typehint is still valid here](https://github.com/DusanKasan/Knapsack/pull/29/files#diff-810cdcfdd8a6b9e1fc0d1e96d7786874R44).

I will have a closer look and merge as soon as possible. Great work!