@issei-m 2017-04-14
@DusanKasan Thank you for quick response.

> except that the callable typehint is still valid here.

I don't understand what it means. `Collection::from()` just looks a static constructor copy, can be declared the same one.