@DusanKasan 2017-04-14
Sorry, nevermind, i misread the PR (thought you removed the callable typehint but you actually added it). :)