@DusanKasan 2017-04-14
Looks good to me, the performance impact is really significant. Thanks for the contribution.